refactor: address dedup audit follow-ups for workers
