+++ workers/espn-client/src/logging.ts
@@ -1,2 +1,10 @@
-export { logEvalEvent } from '@flaim/worker-shared';
-export type { TraceLogEvent } from '@flaim/worker-shared';
+import { logEvalEvent as sharedLogEvalEvent, type TraceLogEvent as SharedTraceLogEvent } from '@flaim/worker-shared';
+
+export type TraceLogEvent = Omit<SharedTraceLogEvent, 'service' | 'phase'> & {
+  service: 'espn-client';
+  phase: 'execute_start' | 'execute_end' | 'execute_error';
+};
+
+export function logEvalEvent(event: TraceLogEvent): void {
+  sharedLogEvalEvent(event);
+}
