+++ workers/shared/package.json
@@ -5,10 +5,12 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "scripts": {
+    "test": "vitest run",
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20241218.0",
-    "typescript": "^5.7.2"
+    "typescript": "^5.7.2",
+    "vitest": "~3.0.0"
   }
 }
