+++ workers/sleeper-client/src/shared/handlers/types.ts
@@ -1,0 +1,13 @@
+import type { Env, ToolParams, ExecuteResponse } from '../../types';
+
+export type HandlerFn = (
+  env: Env,
+  params: ToolParams,
+  authHeader?: string,
+  correlationId?: string
+) => Promise<ExecuteResponse>;
+
+export interface SleeperSportConfig {
+  sport: 'football' | 'basketball';
+  statePath: '/state/nfl' | '/state/nba';
+}
